+++ doc/changelog.md
@@ -1,9 +1,14 @@
-# version 1.5.1
+# version 1.6.0
 - new coin: Monero (XMR)
+- new coin: Nano (XNO)
 - new coin: NEM (XEM)
 - new coin: Zilliqa (ZIL)
     - Legacy address zil... supported
     - New V2 address 0x... supported
+- Litecoin:
+    - Taproot address support ltc1p... (default)
+    - New supporting protocol BIP86
+    - Legacy  address support L...
 - cargo update:
     - egui (0.34.3 -> 0.36.0)
     - eframe (0.34.3 -> 0.36.0)
@@ -34,6 +39,7 @@
 
 
 ---
+
 
 # version 1.4.0
 - Bitcoin: